Allow 'date' class on datepicker input element
The original code assumes that if the element has a 'date' class it must be a container that has an element with the 'add-on' class. In my case that causes problems because I apply the 'date' class to input elements, which cannot have an 'add-on' child. This change is intended to be non-breaking while allowing the date class to be applied to input elements.